initial upgrade to .net9
This required rebuilding the .sln and csproj files.  The silverlight and
setup projects were removed as they are no longer supported.

All dll's were removed except the ICSharpCode.TextEditor.dll which is
version 2 since the only nuget package available is version 3 which
brings breaking changes.

Also, when converting the csproj files to the new SDK format, it was
necessary to add post-build steps to copy the referenced plugin dlls
to the Plugins folder in the output.

Currently, the following issues need to be addressed:

- XmlCompletionProvider no long supports BeginInvoke
- XmlFoldingStrategy throws exception on Stack (was present in .net4)
- App crash when clicking inside ICSharpCode.TextEditorControl