fix binding errors in Snippets and WpfDocumentView
Snippets.xaml contained a Context binding with a property that did not
exist in the DataContext and served no purpose.

WpfDocumentView's grey error overlay on the xaml render control had a
binding on its source that did not work.  However, the control sets
the source in its code behind so the binding was superfluous.