fix find and replace out-of-range exception
If the user performs a find and replace when the caret is at the very
end/bottom of the text editor, an out of range exception can be thrown
if the resulting text in the editor is shortened as a result of the
replacements.  This is because the original index that is attempted to
be reselected no longer exists.  This fixes it by doing a Math.MIn.