fix prior plugin load exception workaround
Build scripts added to the main csproj file that copied the plugin DLLs
incorrectly include the plugin base dll which was reflected when
loading UserControl types.  This would throw exceptions.  This fixes the
build scripts but leaves the check and logs as a warning in case
something similar were to happen again.