initial xml tag edit matching
This is the initial attempt to perform xml tag matching as the user
edits the XAML file.  After each keystroke, the XML is parsed for
mismatched tags.  If there is one and only one found and the editor
caret is withing one of the end, the other end's name will be replaced.

One problem encountered was the Undo and Redo actions.  Brace
matching must be bypassed when the user is invoking either of these.
Otherwise, strange mismatches will happen since the UndoStack
will already handle this.