change temp folder to local appdata
The application was using the current working directory as a temp folder
to save new documents.  However, if the user were to search for, say,
"kaxaml.exe" and launch it in the Start Menu, the current directory
would be the OS System folder.  This would then throw an exception in
the application since it will not have right access (unless launch with
admin rights).

This commit fixes it by using Local AppData as the base folder instead.
Settings are saved in the app.config file.  It also adds a button to the
About screen for easy access.