fix new document temp folder error
Related to 37dfff148e9acae0c80bca0a1571a5566e86da2c.  When the user
creates a new document (e.g. ctrl+T), the error message is displayed
about not being able to save to the systems folder if the exe was
launched by searching the Start menu.  This changes it to use the same
temp folder variable introduced in that commit.