initial version of references plugin
This is the first working version of the references Plugin that allows
the user to manually load external DLLs.  These can then be referenced
directly in the XAML editor.

This is based off of the great work punker76 did:

https://github.com/punker76/kaxaml

Many of the UI and code components have been revised or updated to the
latest language features, but the concept is much the same.

Still needs more shakedown.  Especially when loading DLLs from different
.net versions.  But so far it seems stable.