support assembly references snippet
This allows for a well-formatted XAML comment to have assemblies
automatically when the document is loaded.  This also introduced the
XamDocumentManager to allow dependency injection of currently
loaded and selected documents.  Also, the AssemblyCacheManager
was enhanced to more intelligently load assemblies based on a hash.

Still need to address the new XamlObjectWriter not fully supporting
all XAML.  For example, inline control templates and other XAML that
should not be parsed or attempted to be loaded immediately will
throw exceptions.