Revert to using XamlReader for parsing
This changes the main XAML parsing logic from the XamlXmlReader back
to the older XamlReader orginally used.  While the newer reader seemed
to handle dynamic loading better out of the box, it failed to handle
XAML that contained deferred-loaded content, e.g. Control Templates.

The change was originally made to allow parsing of XAML after assemblies
were loaded via the new AssemblyReferences plugin.  If the parser made
an attempt with namespaces that had not yet been loaded, it would not
just fail but also store the broken reference in the XamlReder's cache.
After this happened, the parser would never succeed even after loading
the missing assembly.

This commit fixes that problem by using an empty XamlTypeMapper
when creating the ParserContext.  This forces the XamlReader to only
use namespaces explicitly declared in the XAML text.