change snippet and editor to di
This allows the Snippets plugin and the SnippetEditor control to be
managed by dependency injection.  This required loosening the assumption
in XmlFoldingStrategy that the XML will always have content since the
Editor will be created on startup without any XAML.  Otherwise,
a XML exception would be thrown (and swallowed) whenever the editors
timer to parse fires.

Also, the original code generated a new SnippetEditor instance whenever
the user invoked editing one of the listed snippets.  However, this
resulted in a significant memory leak as the created Window was never
destroyed after closing using the Cancel or Done buttons.  Even opening
the same snippet generated an entirely new Window.  With the new code,
the editor is created by DI as a singleton so there will only ever be
one instance.

Finally, the strategy behind how the snippets were saved on disk was
altered.  The original code would save user edits directly to the
KaxamlSnippets.xml file in the application folder, e.g.
c:\program files\kaxaml.  Now, on startup, the application copies the
file to the users AppData folder if not already present.  Any edits made
will be saved there to avoid any need for elevated privileges or if
more than one user were to work in the application.  A button was
added to the About screen for easy access.