minimize WpfDocumentView memory leak
There seems to be a known issue with the ICSharp Text Edit control with
memory leaks when inside a WinForms host control.  When a tab is
closed via the close button in the tab control, the WpfDocumentView and
WpfDocument are not garbage collected.  This was noticed when the
folding logic was still being called after the document was closed.

This reduces the impact by deactivating the times and detaching all
event handlers in both controls.  This will have to be revisited with a
possible solution of upgrading to a newer/different control.