change XmlFoldingStrategy to use regex parser
Instead of relying on the XmlTextReader, this changes the folding
parsing to use a simple regex manual parse.  This is to provide greater
control of folding including ignoring namespaces which may not
be loaded and folding self-closing tags, neither of which were
supported with XmlTextReader.

The drawback to this is it is significantly slower.  Based on
benchmarking, approximately 10x slower.  This is, of course,
at the micro second level but it would still be nice to reduce it
since it is called so often.  Will look at TurboXml next.