change fold parsing to TurboXml
Changing to a manual RegEx based parsing when generating folds solved
the problem where foldings failed to generate because the XML was not
fully compliant and well-formed.  For example, snippets were often
technically incomplete and the XmlTextReader would fail.

But this came with a large performance penalty since XmlTextReader is so
well optimized.  To improve this, the TurboXml library was brought in.
It allows much better control over the lower-level parsing and does not
perform validation other then the XML tags must be properly formed.

To confirm performance differences, the Kaxaml.Benchmarks project
was created using the BenchmarkDotNet library.  The different
techniques were tested.  XmlTextReader was the fastest with TurboXml
within the same order of magnitude.  The RegEx was about an order
of magnitude slower.