remove TextDragger automatic copy to clipboard
The Color Picker and Snippet plugins both use the TextDragger control
which automatically copies the bound property to the clipboard on a
single click.  Not quite sure the reason for this but it seemed
unnecessary and, at time, problematic if the user is relying on the
prior clipboard content.  This removes the functionality.