Simplify external plugin library resolution
Have the main csproj file use build steps to copy the external plugin
dlls was still problematic after upgrading to .net 9.  Since those steps
were not guaranteed to run in a particular order, older plugin dlls were
being copied to the Plugins subfolder.  To guarantee the correct dlls
were present required a clean and rebuild.  This removes all custom
logic from the csproj and relies on a simpler explicit dll name
collection in the ApplicationDiServiceProvider.

This also includes the CleanBuildFolders script.  This can be useful
when troubleshooting by guaranteeing all caches are cleared.